Creates a new base test file for search-client package
