+++ search_client/test/cases/__init__.py
@@ -1,0 +1,1 @@
+from search_client.test.cases.integration import BaseSearchClientIntegrationTestCase
